Update Flask plugins and related dependencies
This changeset updates a few of the Flask plugins we rely on to their latest versions, as well as the psycopg2-binary library.  It required a few adjustments to how our tests are configured and run, which are probably due for a larger refactor given how much has changed with these libraries and the anticipated change of going from SQLAlchemy 1.4.x to 2.0.x.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>