Downgrade eventlet to 0.33.3
This changeset reverts a recent Dependabot update for eventlet back to 0.33.3, the last known working version for deployment.  The 0.34.1 release changed a bunch of deployment related things, including an overhaul of the eventlet/__init__.py file that removed the __version__ attribute.  This is checked when the deployment takes place in a Cloud Foundry environment (e.g., cloud.gov), ultimately by something internal within gunicorn.

It is unclear if the breakdown is with gunicorn itself or with the way that the Cloud Foundry Python buildpack operates during deployment, but it seems that either we will have to wait for an eventlet bug fix and/or updates to gunicorn and the Cloud Foundry Python buildack before we can update eventlet itself again.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>