Update try/except block to catch Exception
This changeset updates a try/except block to catch Exception instead of BaseException.  Using BaseException is an anti-pattern and is not intended for user-defined code; it can also lead to other unintended side effects if not handled properly.

This is enough of a concern that the new release of flake8-bugbear now issues warnings when it sees BaseException being caught without an immediate re-raise.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>