Add code review documentation
This changeset adds documentation for what to expect and how to conduct and receive code reviews.  We want to improve our team practices and ensure that all members know how to engage and what to expect with code reviews.  Furthermore, this can serve as another model and positive example of how to collaborate as a team in the broader open source community.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>