Add E2E configuration information to docs
This changeset adds a bit of missing information to our README and documentation for E2E test configuration.  This is required for the project to be set up correctly, and while it will work out of the box, there are a couple of adjustments that need to be made to ensure everything is configured correctly.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>