Adjust properties further to only apply to spaces
The SES and SNS modules are strictly for services that are instantiated within a space, while the egress_space is creating a separate space. The shift of the recursive delete from being on a service to being at the space level means that the SES and SNS modules do not have to track it at all, it should only be handled at the space level.  The same goes for the allow_ssh flag.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>