Revert recent Terraform changes to prod and demo
This changeset reverts a few of our recent changes to the production and demo Terraform files to re-enable deployments.

We are working through fixing this and will re-introduce these changes once we figure out how to get the changes to properly work (which are tied to the infrastructure checks as well).

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>