Bump Redis plan to 5node-large
This changeset boosts our production Redis configuration to the 5node-large plan to give us a bit more wiggle room for heavy production workloads.  This should still keep us in the threshold of the second block of 10 nodes we are already signed up for as we currently have 12 nodes going across all environments at the moment.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>