Remove SocketIO (#2019)
* Remove SocketIO

We have tried a few times to enable SocketIO in our app and it has been a challenge each step of the way, and now there is a dependency incompatibility it seems.  Given that we are in maintenance mode with the applicaiton currently and hve different plans for the interface, we are removing this dependency.

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>

* Remove more unused code and address Flake8 warnings

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>

* Revert changes to fix test

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>

---------

Signed-off-by: Carlo Costino <carlo.costino@gsa.gov>